* added Parallel Colt library. * added GraphicsUtil class which provide some Graphics facilities as drawing multi line text, calculating text bounds... * IcyCanvas:   - fixed minors issues.   - added canvasToImageLogDelta() methods. * IcyCanvas2D:   - added canvasToImageLogDelta() methods.   - added getTransform() & getInverseTransform() methods which return AffineTransform object for fast canvas <--> image Graphics context change. * minor fix in ToolTipFrame. * faster histogram calculation. * added temporary fix for unexpected repaint from Ribbon component. * ROI :  - manipulation reworked.  - improved rendering. * added Hint on focused ROI. * some fixes and tweaks.

+++ icy/gui/sequence/SequenceInfosPanel.java
@@ -196,7 +196,7 @@ public class SequenceInfosPanel extends InspectorSubPanel
             nameLabel.setText(sequence.getName());
             dimensionLabel.setText(sizeX + " x " + sizeY + " x " + sizeZ + " x " + sizeT);
             channelLabel.setText(sizeC + " - " + sequence.getDataType_());
-            sizeLabel.setText(UnitUtil.getBytesString(sizeX * sizeY * sizeZ * sizeT * sizeC
+            sizeLabel.setText(UnitUtil.getBytesString((double) sizeX * (double) sizeY * sizeZ * sizeT * sizeC
                     * sequence.getDataType_().getSize()));
             resXLabel.setText(UnitUtil.getBestUnitInMeters(pxSizeX, 2, UnitPrefix.MILLI));
             resYLabel.setText(UnitUtil.getBestUnitInMeters(pxSizeY, 2, UnitPrefix.MILLI));
